added in 15/8/2024

+++ CPP0507 - TInh Tong 2 Phan So.cpp
@@ -15,7 +15,7 @@
 using namespace std;
 
 struct PhanSo{
-    int tu, mau;
+    ll tu, mau;
 };
 
 void nhap (PhanSo &A){
